+++ src/ebus_sdk/__init__.py
@@ -65,7 +65,7 @@ from .declaration import (
 # MQTT client
 from ebus_mqtt_client import MqttClient
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
 
 __all__ = [
     # Homie classes